+++ matrix-sdk-android/src/main/java/im/vector/matrix/android/internal/auth/db/SessionParamsEntity.kt
@@ -22,5 +22,8 @@ import io.realm.annotations.PrimaryKey
 internal open class SessionParamsEntity(
         @PrimaryKey var userId: String = "",
         var credentialsJson: String = "",
-        var homeServerConnectionConfigJson: String = ""
+        var homeServerConnectionConfigJson: String = "",
+        // Set to false when the token is invalid and the user has been soft logged out
+        // In case of hard logout, this object is deleted from DB
+        var isTokenValid: Boolean = true
 ) : RealmObject()