+++ matrix-sdk-android/src/main/java/im/vector/matrix/android/internal/session/DefaultSession.kt
@@ -23,7 +23,7 @@ import androidx.lifecycle.LiveData
 import dagger.Lazy
 import im.vector.matrix.android.api.MatrixCallback
 import im.vector.matrix.android.api.auth.data.SessionParams
-import im.vector.matrix.android.api.failure.ConsentNotGivenError
+import im.vector.matrix.android.api.failure.GlobalError
 import im.vector.matrix.android.api.pushrules.PushRuleService
 import im.vector.matrix.android.api.session.InitialSyncProgressService
 import im.vector.matrix.android.api.session.Session
@@ -42,12 +42,16 @@ import im.vector.matrix.android.api.session.signout.SignOutService
 import im.vector.matrix.android.api.session.sync.FilterService
 import im.vector.matrix.android.api.session.sync.SyncState
 import im.vector.matrix.android.api.session.user.UserService
+import im.vector.matrix.android.internal.auth.SessionParamsStore
 import im.vector.matrix.android.internal.crypto.DefaultCryptoService
 import im.vector.matrix.android.internal.database.LiveEntityObserver
 import im.vector.matrix.android.internal.session.sync.SyncTaskSequencer
 import im.vector.matrix.android.internal.session.sync.SyncTokenStore
 import im.vector.matrix.android.internal.session.sync.job.SyncThread
 import im.vector.matrix.android.internal.session.sync.job.SyncWorker
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.GlobalScope
+import kotlinx.coroutines.launch
 import org.greenrobot.eventbus.EventBus
 import org.greenrobot.eventbus.Subscribe
 import org.greenrobot.eventbus.ThreadMode
@@ -76,6 +80,7 @@ internal class DefaultSession @Inject constructor(override val sessionParams: Se
                                                   private val contentUrlResolver: ContentUrlResolver,
                                                   private val syncTokenStore: SyncTokenStore,
                                                   private val syncTaskSequencer: SyncTaskSequencer,
+                                                  private val sessionParamsStore: SessionParamsStore,
                                                   private val contentUploadProgressTracker: ContentUploadStateTracker,
                                                   private val initialSyncProgressService: Lazy<InitialSyncProgressService>,
                                                   private val homeServerCapabilitiesService: Lazy<HomeServerCapabilitiesService>)
@@ -97,6 +102,9 @@ internal class DefaultSession @Inject constructor(override val sessionParams: Se
     private var isOpen = false
 
     private var syncThread: SyncThread? = null
+
+    override val isOpenable: Boolean
+        get() = sessionParamsStore.get(myUserId)?.isTokenValid ?: false
 
     @MainThread
     override fun open() {
@@ -169,8 +177,16 @@ internal class DefaultSession @Inject constructor(override val sessionParams: Se
     }
 
     @Subscribe(threadMode = ThreadMode.MAIN)
-    fun onConsentNotGivenError(consentNotGivenError: ConsentNotGivenError) {
-        sessionListeners.dispatchConsentNotGiven(consentNotGivenError)
+    fun onGlobalError(globalError: GlobalError) {
+        if (globalError is GlobalError.InvalidToken
+                && globalError.softLogout) {
+            // Mark the token has invalid
+            GlobalScope.launch(Dispatchers.IO) {
+                sessionParamsStore.setTokenInvalid(myUserId)
+            }
+        }
+
+        sessionListeners.dispatchGlobalError(globalError)
     }
 
     override fun contentUrlResolver() = contentUrlResolver