+++ vector/src/main/java/im/vector/riotx/core/error/ErrorFormatter.kt
@@ -25,14 +25,15 @@ import java.net.SocketTimeoutException
 import java.net.UnknownHostException
 import javax.inject.Inject
 
-class ErrorFormatter @Inject constructor(private val stringProvider: StringProvider) {
+interface ErrorFormatter {
+    fun toHumanReadable(throwable: Throwable?): String
+}
 
-    fun toHumanReadable(failure: Failure): String {
-        // Default
-        return failure.localizedMessage
-    }
+class DefaultErrorFormatter @Inject constructor(
+        private val stringProvider: StringProvider
+) : ErrorFormatter {
 
-    fun toHumanReadable(throwable: Throwable?): String {
+    override fun toHumanReadable(throwable: Throwable?): String {
         return when (throwable) {
             null                         -> null
             is Failure.NetworkConnection -> {
@@ -41,6 +42,7 @@ class ErrorFormatter @Inject constructor(private val stringProvider: StringProvi
                         stringProvider.getString(R.string.error_network_timeout)
                     throwable.ioException is UnknownHostException   ->
                         // Invalid homeserver?
+                        // TODO Check network state, airplane mode, etc.
                         stringProvider.getString(R.string.login_error_unknown_host)
                     else                                            ->
                         stringProvider.getString(R.string.error_no_network)
@@ -52,23 +54,23 @@ class ErrorFormatter @Inject constructor(private val stringProvider: StringProvi
                         // Special case for terms and conditions
                         stringProvider.getString(R.string.error_terms_not_accepted)
                     }
-                    throwable.error.code == MatrixError.FORBIDDEN
+                    throwable.error.code == MatrixError.M_FORBIDDEN
                             && throwable.error.message == "Invalid password" -> {
                         stringProvider.getString(R.string.auth_invalid_login_param)
                     }
-                    throwable.error.code == MatrixError.USER_IN_USE          -> {
+                    throwable.error.code == MatrixError.M_USER_IN_USE        -> {
                         stringProvider.getString(R.string.login_signup_error_user_in_use)
                     }
-                    throwable.error.code == MatrixError.BAD_JSON             -> {
+                    throwable.error.code == MatrixError.M_BAD_JSON           -> {
                         stringProvider.getString(R.string.login_error_bad_json)
                     }
-                    throwable.error.code == MatrixError.NOT_JSON             -> {
+                    throwable.error.code == MatrixError.M_NOT_JSON           -> {
                         stringProvider.getString(R.string.login_error_not_json)
                     }
-                    throwable.error.code == MatrixError.LIMIT_EXCEEDED       -> {
+                    throwable.error.code == MatrixError.M_LIMIT_EXCEEDED     -> {
                         limitExceededError(throwable.error)
                     }
-                    throwable.error.code == MatrixError.THREEPID_NOT_FOUND   -> {
+                    throwable.error.code == MatrixError.M_THREEPID_NOT_FOUND -> {
                         stringProvider.getString(R.string.login_reset_password_error_not_found)
                     }
                     else                                                     -> {