+++ vector/src/main/java/im/vector/riotx/features/MainActivity.kt
@@ -19,9 +19,11 @@ package im.vector.riotx.features
 import android.app.Activity
 import android.content.Intent
 import android.os.Bundle
+import android.os.Parcelable
 import androidx.appcompat.app.AlertDialog
 import com.bumptech.glide.Glide
 import im.vector.matrix.android.api.MatrixCallback
+import im.vector.matrix.android.api.failure.GlobalError
 import im.vector.riotx.R
 import im.vector.riotx.core.di.ActiveSessionHolder
 import im.vector.riotx.core.di.ScreenComponent
@@ -31,6 +33,10 @@ import im.vector.riotx.core.platform.VectorBaseActivity
 import im.vector.riotx.core.utils.deleteAllFiles
 import im.vector.riotx.features.home.HomeActivity
 import im.vector.riotx.features.login.LoginActivity
+import im.vector.riotx.features.notifications.NotificationDrawerManager
+import im.vector.riotx.features.signout.hard.SignedOutActivity
+import im.vector.riotx.features.signout.soft.SoftLogoutActivity
+import kotlinx.android.parcel.Parcelize
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.GlobalScope
 import kotlinx.coroutines.launch
@@ -38,23 +44,37 @@ import kotlinx.coroutines.withContext
 import timber.log.Timber
 import javax.inject.Inject
 
+@Parcelize
+data class MainActivityArgs(
+        val clearCache: Boolean = false,
+        val clearCredentials: Boolean = false,
+        val isUserLoggedOut: Boolean = false,
+        val isSoftLogout: Boolean = false
+) : Parcelable
+
+/**
+ * This is the entry point of RiotX
+ * This Activity, when started with argument, is also doing some cleanup when user disconnects,
+ * clears cache, is logged out, or is soft logged out
+ */
 class MainActivity : VectorBaseActivity() {
 
     companion object {
-        private const val EXTRA_CLEAR_CACHE = "EXTRA_CLEAR_CACHE"
-        private const val EXTRA_CLEAR_CREDENTIALS = "EXTRA_CLEAR_CREDENTIALS"
+        private const val EXTRA_ARGS = "EXTRA_ARGS"
 
         // Special action to clear cache and/or clear credentials
-        fun restartApp(activity: Activity, clearCache: Boolean = false, clearCredentials: Boolean = false) {
+        fun restartApp(activity: Activity, args: MainActivityArgs) {
             val intent = Intent(activity, MainActivity::class.java)
             intent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK or Intent.FLAG_ACTIVITY_CLEAR_TASK)
 
-            intent.putExtra(EXTRA_CLEAR_CACHE, clearCache)
-            intent.putExtra(EXTRA_CLEAR_CREDENTIALS, clearCredentials)
+            intent.putExtra(EXTRA_ARGS, args)
             activity.startActivity(intent)
         }
     }
 
+    private lateinit var args: MainActivityArgs
+
+    @Inject lateinit var notificationDrawerManager: NotificationDrawerManager
     @Inject lateinit var sessionHolder: ActiveSessionHolder
     @Inject lateinit var errorFormatter: ErrorFormatter
 
@@ -64,46 +84,73 @@ class MainActivity : VectorBaseActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        val clearCache = intent.getBooleanExtra(EXTRA_CLEAR_CACHE, false)
-        val clearCredentials = intent.getBooleanExtra(EXTRA_CLEAR_CREDENTIALS, false)
-
+        args = parseArgs()
+        if (args.clearCredentials || args.isUserLoggedOut) {
+            clearNotifications()
+        }
         // Handle some wanted cleanup
-        if (clearCache || clearCredentials) {
-            doCleanUp(clearCache, clearCredentials)
+        if (args.clearCache || args.clearCredentials) {
+            doCleanUp()
         } else {
-            start()
+            startNextActivityAndFinish()
         }
     }
 
-    private fun doCleanUp(clearCache: Boolean, clearCredentials: Boolean) {
+    private fun clearNotifications() {
+        // Dismiss all notifications
+        notificationDrawerManager.clearAllEvents()
+        notificationDrawerManager.persistInfo()
+    }
+
+    private fun parseArgs(): MainActivityArgs {
+        val argsFromIntent: MainActivityArgs? = intent.getParcelableExtra(EXTRA_ARGS)
+        Timber.w("Starting MainActivity with $argsFromIntent")
+
+        return MainActivityArgs(
+                clearCache = argsFromIntent?.clearCache ?: false,
+                clearCredentials = argsFromIntent?.clearCredentials ?: false,
+                isUserLoggedOut = argsFromIntent?.isUserLoggedOut ?: false,
+                isSoftLogout = argsFromIntent?.isSoftLogout ?: false
+        )
+    }
+
+    private fun doCleanUp() {
         val session = sessionHolder.getSafeActiveSession()
         if (session == null) {
-            start()
+            startNextActivityAndFinish()
             return
         }
         when {
-            clearCredentials -> session.signOut(object : MatrixCallback<Unit> {
-                override fun onSuccess(data: Unit) {
-                    Timber.w("SIGN_OUT: success, start app")
-                    sessionHolder.clearActiveSession()
-                    doLocalCleanupAndStart()
-                }
+            args.clearCredentials -> session.signOut(
+                    !args.isUserLoggedOut,
+                    object : MatrixCallback<Unit> {
+                        override fun onSuccess(data: Unit) {
+                            Timber.w("SIGN_OUT: success, start app")
+                            sessionHolder.clearActiveSession()
+                            doLocalCleanupAndStart()
+                        }
 
-                override fun onFailure(failure: Throwable) {
-                    displayError(failure, clearCache, clearCredentials)
-                }
-            })
-            clearCache       -> session.clearCache(object : MatrixCallback<Unit> {
-                override fun onSuccess(data: Unit) {
-                    session.startSyncing(applicationContext)
-                    doLocalCleanupAndStart()
-                }
+                        override fun onFailure(failure: Throwable) {
+                            displayError(failure)
+                        }
+                    })
+            args.clearCache       -> session.clearCache(
+                    object : MatrixCallback<Unit> {
+                        override fun onSuccess(data: Unit) {
+                            session.startSyncing(applicationContext)
+                            doLocalCleanupAndStart()
+                        }
 
-                override fun onFailure(failure: Throwable) {
-                    displayError(failure, clearCache, clearCredentials)
-                }
-            })
+                        override fun onFailure(failure: Throwable) {
+                            displayError(failure)
+                        }
+                    })
         }
+    }
+
+    override fun handleInvalidToken(globalError: GlobalError.InvalidToken) {
+        // No op here
+        Timber.w("Ignoring invalid token global error")
     }
 
     private fun doLocalCleanupAndStart() {
@@ -119,24 +166,43 @@ class MainActivity : VectorBaseActivity() {
             }
         }
 
-        start()
+        startNextActivityAndFinish()
     }
 
-    private fun displayError(failure: Throwable, clearCache: Boolean, clearCredentials: Boolean) {
+    private fun displayError(failure: Throwable) {
         AlertDialog.Builder(this)
                 .setTitle(R.string.dialog_title_error)
                 .setMessage(errorFormatter.toHumanReadable(failure))
-                .setPositiveButton(R.string.global_retry) { _, _ -> doCleanUp(clearCache, clearCredentials) }
-                .setNegativeButton(R.string.cancel) { _, _ -> start() }
+                .setPositiveButton(R.string.global_retry) { _, _ -> doCleanUp() }
+                .setNegativeButton(R.string.cancel) { _, _ -> startNextActivityAndFinish() }
                 .setCancelable(false)
                 .show()
     }
 
-    private fun start() {
-        val intent = if (sessionHolder.hasActiveSession()) {
-            HomeActivity.newIntent(this)
-        } else {
-            LoginActivity.newIntent(this, null)
+    private fun startNextActivityAndFinish() {
+        val intent = when {
+            args.clearCredentials
+            && !args.isUserLoggedOut         ->
+                // User has explicitly asked to log out
+                LoginActivity.newIntent(this, null)
+            args.isSoftLogout                ->
+                // The homeserver has invalidated the token, with a soft logout
+                SoftLogoutActivity.newIntent(this)
+            args.isUserLoggedOut             ->
+                // the homeserver has invalidated the token (password changed, device deleted, other security reason
+                SignedOutActivity.newIntent(this)
+            sessionHolder.hasActiveSession() ->
+                // We have a session.
+                // Check it can be opened
+                if (sessionHolder.getActiveSession().isOpenable) {
+                    HomeActivity.newIntent(this)
+                } else {
+                    // The token is still invalid
+                    SoftLogoutActivity.newIntent(this)
+                }
+            else                             ->
+                // First start, or no active session
+                LoginActivity.newIntent(this, null)
         }
         startActivity(intent)
         finish()