+++ vector/src/main/java/im/vector/riotx/features/home/room/detail/timeline/action/MessageActionsEpoxyController.kt
@@ -44,9 +44,7 @@ class MessageActionsEpoxyController @Inject constructor(private val stringProvid
             bottomSheetMessagePreviewItem {
                 id("preview")
                 avatarRenderer(avatarRenderer)
-                avatarUrl(state.informationData.avatarUrl ?: "")
-                senderId(state.informationData.senderId)
-                senderName(state.senderName())
+                matrixItem(state.informationData.matrixItem)
                 movementMethod(createLinkMovementMethod(listener))
                 body(body.linkify(listener))
                 time(state.time())