+++ vector/src/main/java/im/vector/riotx/features/login/LoginViewModel.kt
@@ -17,7 +17,14 @@
 package im.vector.riotx.features.login
 
 import android.content.Context
-import com.airbnb.mvrx.*
+import androidx.fragment.app.FragmentActivity
+import com.airbnb.mvrx.ActivityViewModelContext
+import com.airbnb.mvrx.Fail
+import com.airbnb.mvrx.Loading
+import com.airbnb.mvrx.MvRxViewModelFactory
+import com.airbnb.mvrx.Success
+import com.airbnb.mvrx.Uninitialized
+import com.airbnb.mvrx.ViewModelContext
 import com.squareup.inject.assisted.Assisted
 import com.squareup.inject.assisted.AssistedInject
 import im.vector.matrix.android.api.MatrixCallback
@@ -38,6 +45,7 @@ import im.vector.riotx.core.utils.DataSource
 import im.vector.riotx.core.utils.PublishDataSource
 import im.vector.riotx.features.notifications.PushRuleTriggerListener
 import im.vector.riotx.features.session.SessionListener
+import im.vector.riotx.features.signout.soft.SoftLogoutActivity
 import timber.log.Timber
 import java.util.concurrent.CancellationException
 
@@ -62,8 +70,11 @@ class LoginViewModel @AssistedInject constructor(@Assisted initialState: LoginVi
 
         @JvmStatic
         override fun create(viewModelContext: ViewModelContext, state: LoginViewState): LoginViewModel? {
-            val activity: LoginActivity = (viewModelContext as ActivityViewModelContext).activity()
-            return activity.loginViewModelFactory.create(state)
+            return when (val activity: FragmentActivity = (viewModelContext as ActivityViewModelContext).activity()) {
+                is LoginActivity      -> activity.loginViewModelFactory.create(state)
+                is SoftLogoutActivity -> activity.loginViewModelFactory.create(state)
+                else                  -> error("Invalid Activity")
+            }
         }
     }
 
@@ -99,6 +110,18 @@ class LoginViewModel @AssistedInject constructor(@Assisted initialState: LoginVi
             is LoginAction.ResetPasswordMailConfirmed -> handleResetPasswordMailConfirmed()
             is LoginAction.RegisterAction             -> handleRegisterAction(action)
             is LoginAction.ResetAction                -> handleResetAction(action)
+            is LoginAction.SetupSsoForSessionRecovery -> handleSetupSsoForSessionRecovery(action)
+        }
+    }
+
+    private fun handleSetupSsoForSessionRecovery(action: LoginAction.SetupSsoForSessionRecovery) {
+        setState {
+            copy(
+                    signMode = SignMode.SignIn,
+                    loginMode = LoginMode.Sso,
+                    homeServerUrl = action.homeServerUrl,
+                    deviceId = action.deviceId
+            )
         }
     }
 
@@ -454,7 +477,7 @@ class LoginViewModel @AssistedInject constructor(@Assisted initialState: LoginVi
     }
 
     private fun startAuthenticationFlow() {
-        // No op
+        // Ensure Wizard is ready
         loginWizard
     }